requirements.txt
fastapi
uvicorn
numpy
joblib
scikit-learn
xgboost
scapy
speedtest-cli